fixed a bug in RAS plugin when loading 8-bit palettized images with less than 256 colors

+++ FreeImage/Source/FreeImage/PluginRAS.cpp
@@ -297,7 +297,7 @@ Load(FreeImageIO *io, fi_handle handle, int page, int flags, void *data) {
 					// Read SUN raster colormap
 
 					int numcolors = 1 << header.depth;
-					if(3 * numcolors > header.maplength) {
+					if((DWORD)(3 * numcolors) > header.maplength) {
 						// some RAS may have less colors than the full palette
 						numcolors = header.maplength / 3;
 					} else {
